html+bootstrap #5

+++ src/main/java/stu/cn/ua/service/PassengerService.java
@@ -1,6 +1,7 @@
 package stu.cn.ua.service;
 
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 import stu.cn.ua.domain.Passenger;
 import stu.cn.ua.domain.Ticket;
 import stu.cn.ua.persistence.PassengerRepository;
@@ -23,14 +24,18 @@ public class PassengerService {
     public Passenger save(Passenger passenger){
         return passengerRepository.save(passenger);
     }
-    public void deleteById(Long id){
-    Passenger passenger = passengerRepository.findPassengerByPassengerId(id);
-    Set<Ticket> tickets = passenger.getTickets();
 
-    for(Ticket ticket: tickets){
-    ticketRepository.deleteByTicketId(ticket.getTicketId());
+    @Transactional
+    public void deleteByPassengerId(Long id){
+    Passenger passenger = passengerRepository.findPassengerByPassengerId(id);
+
+    Set<Ticket> tickets = passenger.getTickets();
+if(tickets != null) {
+    for (Ticket ticket : tickets) {
+        ticketRepository.deleteByTicketId(ticket.getTicketId());
     }
     passenger.getTickets().clear();
+}
     passengerRepository.deleteById(passenger.getPassengerId());
     }
